Django-TodoListRest/Django-TodoListRestApi-main/DjangoTodoListRestApi/todo/resources/serializers.py
from rest_framework import serializers

from todo.models import Todo


class TodoSerializer(serializers.ModelSerializer):
    user = serializers.CharField(source='user.username', read_only=True)

    class Meta:
        model = Todo
        fields = '__all__'
